+++ applications/fishing-map/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @globalfishingwatchapp/fishing-map
 
+## 0.28.31
+
+### Patch Changes
+
+- Updated dependencies [0a755cc4]
+  - @globalfishingwatch/layer-composer@5.2.4
+  - @globalfishingwatch/ui-components@2.18.9
+  - @globalfishingwatch/react-hooks@6.0.9
+
 ## 0.28.30
 
 ### Patch Changes