+++ applications/temporalgrid-demo/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @globalfishingwatchapp/temporalgrid-demo
 
+## 0.7.67
+
+### Patch Changes
+
+- Updated dependencies [0a755cc4]
+  - @globalfishingwatch/layer-composer@5.2.4
+  - @globalfishingwatch/react-hooks@6.0.9
+
 ## 0.7.66
 
 ### Patch Changes